send messages from mg to telegram, minor fixes
