+++ Jenkinsfile
@@ -15,15 +15,15 @@ pipeline {
     stages {
         stage('Prepare') {
             steps {
-                sh 'cp config_test.yml.dist config_test.yml'
-                compose 'up -d --build postgres'
-                compose 'run --rm mg_telegram make migrate_test'
+                sh 'cp config_test.yml.dist config.yml'
+                compose 'up -d --build postgres_test'
+                compose 'run --rm mg_telegram_test make migrate_test'
             }
         }
 
         stage('Tests') {
             steps {
-               compose 'run --rm --no-deps mg_telegram make jenkins_test'
+               compose 'run --rm --no-deps mg_telegram_test make jenkins_test'
             }
 
             post {
